% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addcoord.R
\name{addcoord}
\alias{addcoord}
\title{extract coordination from geometry column to be  X and Y the dataframe varible columns}
\usage{
addcoord(site_sf)
}
\arguments{
\item{site_sf}{a sf object of sites}
}
\description{
extract coordination from geometry column to be  X and Y the dataframe varible columns
}
