% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMouthPoint.R
\name{addMouthPoint}
\alias{addMouthPoint}
\title{Add Mount point to site location datafarme}
\usage{
addMouthPoint(df, long = -90.199402, lat = 38.627003, buff = 1)
}
\arguments{
\item{df}{a dataframe containing sit information with columns: id, long, lat}

\item{long}{the longitude of the adding Mount point}

\item{lat}{the latitude of the adding Mount point}

\item{buff}{if buff = 1, add one point;
if buff = 2, then add 9 point around the given long & lat.}
}
\description{
Add Mount point to site location datafarme
}
