% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc2sf.R
\name{arc2sf}
\alias{arc2sf}
\title{convert arc.set to sf}
\usage{
arc2sf(arc.set, riverVert)
}
\arguments{
\item{arc.set}{the arcPair object from makearc contain two columns: from, to}

\item{riverVert}{a sf class object contain columns: id seg vert, X, Y, geometry}
}
\description{
convert arc.set to sf
}
