% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc2vector.R
\name{arc2vector}
\alias{arc2vector}
\title{Convert arc.set to vector.set}
\usage{
arc2vector(arc.set, riverVert)
}
\arguments{
\item{arc.set}{the arcPair object from makearc contain two columns: from, to}

\item{riverVert}{a sf class object contain columns: id seg vert, X, Y, geometry}
}
\description{
Convert arc.set to vector.set
}
