% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bufferintersection.R
\name{bufferintersection}
\alias{bufferintersection}
\title{geospatial operation of intersection between the buffered river centerline sf and site sf}
\usage{
bufferintersection(site_sf, rivers, buffer_dist = 1000)
}
\arguments{
\item{site_sf}{a sf object of sites}

\item{rivers}{a sf object of river centerline}

\item{buffer_dist}{the distance of buffering riversn centerline in meter}
}
\description{
geospatial operation of intersection between the buffered river centerline sf and site sf
}
