% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkbetweenseg.R
\name{linkbetweenseg}
\alias{linkbetweenseg}
\title{connect to the next downstream site across segment}
\usage{
linkbetweenseg(segvert.set, arc_sf.set)
}
\arguments{
\item{segvert.set}{a segvert.set compute by site2segvert() containing column: seg, vert, seg1, vert1, onnected.
Variable seg and vert is the snaped point from site to tth river, whereas column seg1, vert1 are the downstream snaped point}

\item{arc_sf.set}{a arc sf object from makearc() containing columns:
from, seg0, vert0, x0, y0, to, seg1, vert1, x1, y1, geometry}
}
\description{
connect to the next downstream site across segment
}
