% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkwithinseg.R
\name{linkwithinseg}
\alias{linkwithinseg}
\title{connect to the next downstream site within segment}
\usage{
linkwithinseg(site, site_sf, segvert.set = NULL)
}
\arguments{
\item{site}{NULL in the non-first run. In the first run, site is the datafram contain two columns: seg, vert resulted from reorderbyseg()}

\item{site_sf}{a site sf object with column: id, lon, lat, geometry, X, Y, and the snapped point information from the riverdist::xy2segvert() columns: seg, vert, snapdist}

\item{segvert.set}{Null in the first run to identify any withinsegment downstream point. when not NULL, a segvert.set is a dataframe calculate by linkbetweenseg() contains column: seg, vert, seg1, vert1, onnected.
Variable seg and vert is the snaped point from site to tth river, whereas column seg1, vert1 are the downstream snaped point}
}
\description{
connect to the next downstream site within segment
}
