% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makearc.R
\name{makearc}
\alias{makearc}
\title{Make arc pair from riverVert, rivers_net}
\usage{
makearc(riverVert, rivers_net, prefix = "")
}
\arguments{
\item{riverVert}{a sf class object contain columns: id seg vert, X, Y, geometry}

\item{rivers_net}{a riverdist::rivernetwork class object after clean up by riverdist::clearnup()}

\item{prefix}{add text before all site id (defaul is "")}
}
\description{
Make arc pair from riverVert, rivers_net
}
