% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderbyseg.R
\name{reorderbyseg}
\alias{reorderbyseg}
\title{reorder site with grouping by seg in closet order from endvert of each seg}
\usage{
reorderbyseg(site_sf, arc_sf.set)
}
\arguments{
\item{site_sf}{a site sf object with column: id, lon, lat, geometry, X, Y, and the snapped point information from the riverdist::xy2segvert() columns: seg, vert, snapdist}

\item{arc_sf.set}{a arc sf object from makearc() containing columns:
from, seg0, vert0, x0, y0, to, seg1, vert1, x1, y1, geometry}
}
\description{
reorder site with grouping by seg in closet order from endvert of each seg
}
