% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rivernetwork2riverVert.R
\name{rivernetwork2riverVert}
\alias{rivernetwork2riverVert}
\title{convert rivers_net to site_sf}
\usage{
rivernetwork2riverVert(rivers_net, crs = 2163)
}
\arguments{
\item{rivers_net}{a rivernetwork class object from riverdist package.}

\item{crs}{a coordinate reference system for Map projections attempt to portray
the surface of the earth or a portion of the earth on a flat piece of paper or
computer screen.}
}
\description{
convert rivers_net to site_sf
}
