% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segvert2vector.R
\name{segvert2vector}
\alias{segvert2vector}
\title{convert the segvert.set to vector.set}
\usage{
segvert2vector(segvert.set, site_sf)
}
\arguments{
\item{segvert.set}{a segvert.set compute by site2segvert() containing column: seg, vert, seg1, vert1, onnected.
Variable seg and vert is the snaped point from site to tth river, whereas column seg1, vert1 are the downstream snaped point}

\item{site_sf}{a site sf (or dataframe) object that must contain column: id, X, Y, and the snapped point information from the riverdist::xy2segvert() columns: seg, vert}
}
\description{
convert the segvert.set to vector.set
}
