% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifybyselectsite.R
\name{simplifybyselectsite}
\alias{simplifybyselectsite}
\title{Simplify the river topo (arc.set) with sites along the river}
\usage{
simplifybyselectsite(site_sf, arc_sf.set, mouthID)
}
\arguments{
\item{site_sf}{a site sf object with column: id, lon, lat, geometry, X, Y, and the snapped point information from the riverdist::xy2segvert() columns: seg, vert, snapdist}

\item{arc_sf.set}{a arc sf object from makearc() containing columns:
from, seg0, vert0, x0, y0, to, seg1, vert1, x1, y1, geometry}

\item{mouthID}{a string of the site id as the mouth point appear in site_sf}
}
\description{
Simplify the river topo (arc.set) with sites along the river
}
