% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector2sf.R
\name{vector2sf}
\alias{vector2sf}
\title{convert vector.set to sf.set}
\usage{
vector2sf(vector.set, crs)
}
\arguments{
\item{vector.set}{a dataframe from segvert2vector() or dataframe must contain columns: from, seg0, vert0, x0, y0 to, seg1, vert1, x1, y1}

\item{crs}{a coordinate reference system for Map projections attempt to portray
the surface of the earth or a portion of the earth on a flat piece of paper or
computer screen.}
}
\description{
convert vector.set to sf.set
}
